Provide optional Consul agent co-process
Services register health checks with a specific Consul agent, and this
makes health checks fail if we try to treat the cluster as a single
Consul as we do in Triton deployments or PaaS-like environments.
ref https://github.com/joyent/containerpilot/issues/162

Adding a Consul agent as a co-process with the ContainerPilot 2.3.0
coprocesses feature allows us to have the container join Consul and
get/set health updates as part of the LAN gossip.

- bumps ContainerPilot version to 2.3.0.
- uses templating so that Consul co-process is optional